fix: moves FormProvider up to share state between form tabs
