+++ src/pages/EditPage/EditPage.tsx
@@ -1,8 +1,11 @@
+import { FormProvider, useForm } from "react-hook-form";
 import { IconLink, IconProfileDetailsHeader, LogoDevlinksLarge } from "../../assets";
 import { Button } from "../../components/Button";
 import { Card } from "../../components/Card";
 import Icon from "../../components/Icon/Icon";
 import { Tab } from "../../components/Tab";
+import { Link } from "../../models/Links";
+import { useGetLinks } from "../../services/LinksService/useGetLinks";
 import { LinksTab } from "../LinksTab/LinksTab";
 import { ProfileDetailsTab } from "../ProfileDetailsTab/ProfileDetailsTab";
 import {
@@ -12,35 +15,50 @@ import {
   tabItemStyle,
 } from "./EditPage.css";
 
+type LinkForm = {
+  links: Link[];
+  profilePicture: FileList | null;
+  firstName: string;
+  lastName: string;
+  email: string;
+};
+
 export const EditPage = () => {
+  const { data: linksData } = useGetLinks();
+  const methods = useForm<LinkForm>({
+    defaultValues: { links: linksData },
+  });
+
   return (
-    <Tab.Root defaultValue="links">
-      <Card className={appHeader}>
-        <LogoDevlinksLarge className={devLinksLogo} />
-        <Tab.List>
-          <Tab.Item value="links" className={tabItemStyle}>
-            <Icon icon={IconLink} />
-            Links
-          </Tab.Item>
-          <Tab.Item
-            value="profileDetails"
-            className={tabItemStyle}
-            width="187px"
-          >
-            <Icon icon={IconProfileDetailsHeader} />
-            Profile Details
-          </Tab.Item>
-        </Tab.List>
-        <Button variant="secondary" width="114px">
-          Preview
-        </Button>
-      </Card>
-      <Tab.Content value="links" className={tabContentStyle}>
-        <LinksTab />
-      </Tab.Content>
-      <Tab.Content value="profileDetails" className={tabContentStyle}>
-        <ProfileDetailsTab />
-      </Tab.Content>
-    </Tab.Root>
+    <FormProvider {...methods}>
+      <Tab.Root defaultValue="links">
+        <Card className={appHeader}>
+          <LogoDevlinksLarge className={devLinksLogo} />
+          <Tab.List>
+            <Tab.Item value="links" className={tabItemStyle}>
+              <Icon icon={IconLink} />
+              Links
+            </Tab.Item>
+            <Tab.Item
+              value="profileDetails"
+              className={tabItemStyle}
+              width="187px"
+            >
+              <Icon icon={IconProfileDetailsHeader} />
+              Profile Details
+            </Tab.Item>
+          </Tab.List>
+          <Button variant="secondary" width="114px">
+            Preview
+          </Button>
+        </Card>
+        <Tab.Content value="links" className={tabContentStyle}>
+          <LinksTab />
+        </Tab.Content>
+        <Tab.Content value="profileDetails" className={tabContentStyle}>
+          <ProfileDetailsTab />
+        </Tab.Content>
+      </Tab.Root>
+    </FormProvider>
   );
 };
