+++ src/pages/LinksTab/LinksTab.tsx
@@ -1,32 +1,24 @@
 import { useEffect } from "react";
-import { FormProvider, useForm } from "react-hook-form";
-import { Link } from "../../models/Links";
+import { useFormContext } from "react-hook-form";
 import { useGetLinks } from "../../services/LinksService/useGetLinks";
 import { linksTabContainer } from "./LinksTab.css";
 import { CustomizeLinksCard } from "./components/CustomizeLinksCard";
 import { PreviewCard } from "./components/PreviewCard";
 
-type LinkForm = {
-  links: Link[];
-};
 
 export const LinksTab = () => {
   const { data: linksData } = useGetLinks();
-  const methods = useForm<LinkForm>({
-    defaultValues: { links: linksData },
-  });
-  const { reset } = methods;
+  const { reset } = useFormContext();
 
   useEffect(() => {
     reset({ links: linksData });
   }, [linksData, reset]);
 
   return (
-    <FormProvider {...methods}>
-      <div className={linksTabContainer}>
-        <PreviewCard />
-        {linksData && <CustomizeLinksCard />}
-      </div>
-    </FormProvider>
+
+    <div className={linksTabContainer}>
+      <PreviewCard />
+      {linksData && <CustomizeLinksCard />}
+    </div>
   );
 };
